backend/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
langchain
langchain-google-vertexai
python-dotenv
pandas
elevenlabs
